+++ gherkin-utils/javascript/scripts/gherkin-utils.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 #
 
-set -eu -o pipefail
+set -eu -o pipefail -o noglob
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 ${DIR}/../../../node_modules/.bin/ts-node \