+++ gherkin-utils/javascript/src/cli.ts
@@ -6,11 +6,17 @@ const program = new Command()
 program.version(version)
 
 program
-  .command('format <source> <destination>')
-  .description('format <source> and write it to <destination>', {
-    source: 'a file path or a quoted glob, for example features/hello.feature or "features/**/*.feature"',
-    destination: 'a file path or a quoted glob, for example features/hello.feature.md or "features/**/*.feature.md"'
-  })
+  .command('format')
+  .arguments('<source> [destination]')
+  .description(
+    'Formats one or more <source> files and writes them to [destination].\nIf [destination] is not specified, the source files are modified in-place',
+    {
+      source:
+        'A file path or a QUOTED glob, for example features/hello.feature or "features/**/*.feature"',
+      destination:
+        'A file path or a QUOTED glob, for example features/hello.feature.md or "features/**/*.feature.md"',
+    }
+  )
   .action(formatCommand)
 
 program.parse(process.argv)